모든 테스트간 발견된 사항처리

+++ MAMBrowser/BLL/PrivateFileBLL.cs
@@ -121,9 +121,12 @@ namespace MAMBrowser.BLL
             foreach (var seq in seqList)
             {
                 var fileData = Get(seq);
-                if (File.Exists(fileData.FilePath))
+                if (fileData != null)
                 {
-                    _fileProtocol.Delete(fileData.FilePath);
+                    if (File.Exists(fileData.FilePath))
+                    {
+                        _fileProtocol.Delete(fileData.FilePath);
+                    }
                 }
             }
